Basic support for Complex Columns (Access 2007+)
The major new feature in Access 2007 where so called complex columns. These can be multivalued fields, attachments, or MEMO fields with version history. All of them are implemented using hidden helper tables, and the column itself has type MDB_COMPLEX (0x12) which is basically a long int used as a key.

This commit adds basic support for this complex type to libmdb.